creation decorateur currentuser && refactor controller pmscan

+++ src/pmscan/pmscan.service.ts
@@ -1,9 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePmscanDto } from './dto/create-pmscan.dto';
 import { UpdatePmscanDto } from './dto/update-pmscan.dto';
 import { PmscanRepository } from './pmscan.repository';
 import { UsersService } from '../users/users.service';
-import { Prisma } from '@prisma/client';
+import { PMScan, Prisma } from '@prisma/client';
 
 @Injectable()
 export class PmscanService {
@@ -12,18 +16,17 @@ export class PmscanService {
     private readonly usersService: UsersService,
   ) {}
 
-  async create(createPmscanDto: CreatePmscanDto) {
-    const user = await this.usersService.findOne(createPmscanDto.userId);
+  async create(createPmscanDto: CreatePmscanDto, userId: number) {
+    const user = await this.usersService.findOne(userId);
     if (!user) {
       throw new NotFoundException('User not found');
     }
-
     return this.pmscanRepository.create({
       name: createPmscanDto.name,
       deviceId: createPmscanDto.deviceId,
       deviceName: createPmscanDto.deviceName,
       display: Buffer.from(createPmscanDto.display, 'base64'),
-      user: { connect: { id: createPmscanDto.userId } },
+      user: { connect: { id: userId } },
     });
   }
 
@@ -35,19 +38,21 @@ export class PmscanService {
     return this.pmscanRepository.findAllFromUser(userId);
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, userId: number) {
     const pmscan = await this.pmscanRepository.findOne(id);
     if (!pmscan) {
       throw new NotFoundException('PMScan not found');
     }
+    this.checkUserOwnership(pmscan, userId);
     return pmscan;
   }
 
-  async update(id: number, updatePmscanDto: UpdatePmscanDto) {
+  async update(id: number, updatePmscanDto: UpdatePmscanDto, userId: number) {
     const existingPmscan = await this.pmscanRepository.findOne(id);
     if (!existingPmscan) {
-      throw new NotFoundException('PMScan non trouvé');
+      throw new NotFoundException('PMScan not found');
     }
+    this.checkUserOwnership(existingPmscan, userId);
 
     const updateData: Prisma.PMScanUpdateInput = {};
 
@@ -67,8 +72,19 @@ export class PmscanService {
     return this.pmscanRepository.update(id, updateData);
   }
 
-  async remove(id: number) {
+  async remove(id: number, userId: number) {
+    const existingPmscan = await this.pmscanRepository.findOne(id);
+    if (!existingPmscan) {
+      throw new NotFoundException('PMScan non trouvé');
+    }
+    this.checkUserOwnership(existingPmscan, userId);
     await this.pmscanRepository.delete(id);
     return { message: 'PMScan supprimé avec succès' };
+  }
+
+  private async checkUserOwnership(pmscan: PMScan, userId: number) {
+    if (pmscan.userId !== userId) {
+      throw new ForbiddenException('You are not allowed to access this PMScan');
+    }
   }
 }
